+++ packages/@expo/cli/src/prebuild/updateFromTemplate.ts
@@ -69,10 +69,7 @@ export async function updateFromTemplateAsync(
   return {
     hasNewProjectFiles: !!copiedPaths.length,
     // If the iOS folder changes or new packages are added, we should rerun pod install.
-    needsPodInstall:
-      copiedPaths.includes('ios') ||
-      depsResults.hasNewDependencies ||
-      depsResults.hasNewDevDependencies,
+    needsPodInstall: copiedPaths.includes('ios') || !!depsResults.changedDependencies.length,
     ...depsResults,
   };
 }